requirements.txt
requests
PyYAML
jira
